Rename publisher.rs to crate_handler.rs
