+++ cargo-registry/src/crate_handler.rs
